requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-jazzmin>=2.6
Pillow>=10.0
django-filter>=23.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
